Refactor vector extensions and add two methods.
